Add KeyVault without building a config
